fix: wire cmd_doctor dispatch, enable branch coverage, fix coverage parsing
- Add cmd_doctor import and kwarg to main() in sm.py (fixes KeyError)
- Enable branch = true in pyproject.toml [tool.coverage.run]
- Fix _parse_coverage and _parse_missing_lines regexes to handle
  optional Branch/BrPart columns in coverage report output
- Add dispatch tests for all 4 missing verbs (doctor, sail, refit, init)
- Extract TestDetectProjectType to test_detection.py (code-sprawl fix)
- Add tests for branch-format coverage parsing
