+++ slopmop/sm.py
@@ -713,6 +713,7 @@ def main(args: Optional[List[str]] = None) -> int:
         cmd_buff,
         cmd_commit_hooks,
         cmd_config,
+        cmd_doctor,
         cmd_help,
         cmd_init,
         cmd_refit,
@@ -743,6 +744,7 @@ def main(args: Optional[List[str]] = None) -> int:
             cmd_sail=cmd_sail,
             cmd_refit=cmd_refit,
             cmd_status=cmd_status,
+            cmd_doctor=cmd_doctor,
             cmd_config=cmd_config,
             cmd_help=cmd_help,
             cmd_init=cmd_init,
